+++ src/main/java/com/island/engine/SimulationContext.java
@@ -1,7 +1,6 @@
 package com.island.engine;
 
 import com.island.util.RandomProvider;
-import com.island.nature.view.SimulationView;
 
 /**
  * Context that holds all major components of a running simulation.
@@ -11,11 +10,11 @@ import com.island.nature.view.SimulationView;
 public class SimulationContext<T extends Mortal> {
     private final SimulationWorld<T> world;
     private final GameLoop<T> gameLoop;
-    private final SimulationView view;
+    private final SimulationRenderer view;
     private final RandomProvider random;
 
     public SimulationContext(SimulationWorld<T> world, GameLoop<T> gameLoop, 
-                             SimulationView view, RandomProvider random) {
+                             SimulationRenderer view, RandomProvider random) {
         this.world = world;
         this.gameLoop = gameLoop;
         this.view = view;
@@ -30,16 +29,11 @@ public class SimulationContext<T extends Mortal> {
         return gameLoop;
     }
 
-    public SimulationView getView() {
+    public SimulationRenderer getView() {
         return view;
     }
 
     public RandomProvider getRandom() {
         return random;
-    }
-
-    @Deprecated
-    public SimulationView getConsoleView() {
-        return view;
     }
 }