+++ src/main/java/com/island/engine/WorldListener.java
@@ -1,0 +1,7 @@
+package com.island.engine;
+
+public interface WorldListener {
+    void onEntityAdded(Object key);
+
+    void onEntityRemoved(Object key);
+}